Fix compilation warnings.

+++ src/voip/audiostream.c
@@ -618,8 +618,8 @@ int audio_stream_start_full(AudioStream *stream, RtpProfile *profile, const char
 		stream->dtmfgen=ms_filter_new(MS_DTMF_GEN_ID);
 	else
 		stream->dtmfgen=NULL;
-	rtp_session_signal_connect(rtps,"telephone-event",(RtpCallback)on_dtmf_received,(unsigned long)stream);
-	rtp_session_signal_connect(rtps,"payload_type_changed",(RtpCallback)mediastream_payload_type_changed,(unsigned long)&stream->ms);
+	rtp_session_signal_connect(rtps,"telephone-event",(RtpCallback)on_dtmf_received,stream);
+	rtp_session_signal_connect(rtps,"payload_type_changed",(RtpCallback)mediastream_payload_type_changed,&stream->ms);
 	
 	if (stream->ms.state==MSStreamPreparing){
 		/*we were using the dummy preload graph, destroy it but keep sound filters unless no soundcard is given*/
